Fix: Split PatchBuilder test file
