+++ test/unit/domain/services/PatchBuilder.test.ts
@@ -588,787 +588,4 @@ describe('PatchBuilder', () => {
       expect(builder.ops).toHaveLength(6);
     });
   });
-
-  describe('commit()', () => {
-    it('commits a patch and returns the commit SHA', async () => {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      const sha = await builder.commit();
-
-      expect(sha).toBe('c'.repeat(40));
-      expect(persistence.writeBlob).toHaveBeenCalledOnce();
-      expect(persistence.writeTree).toHaveBeenCalledOnce();
-      expect(persistence.commitNodeWithTree).toHaveBeenCalledOnce();
-      expect(persistence.compareAndSwapRef).toHaveBeenCalledWith(
-        'refs/warp/test-graph/writers/writer1',
-        'c'.repeat(40),
-        null,
-      );
-    });
-
-    it('throws error for empty patch', async () => {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      await expect(builder.commit()).rejects.toThrow('Cannot commit empty patch');
-    });
-
-    it('creates commit with schema:2 in trailers', async () => {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      await builder.commit();
-
-      // Check the commit message passed to commitNodeWithTree
-      const commitCall = persistence.commitNodeWithTree.mock.calls[0]![0];
-      const decoded = decodePatchMessage(commitCall.message);
-
-      expect(decoded.schema).toBe(2);
-      expect(decoded.writer).toBe('writer1');
-      expect(decoded.graph).toBe('test-graph');
-      expect(decoded.lamport).toBe(1);
-    });
-
-    it('increments lamport when continuing from existing ref', async () => {
-      const persistence = createMockPersistence();
-      const existingSha = 'd'.repeat(40);
-      const existingPatchOid = 'e'.repeat(40);
-      // Simulate existing ref with lamport 5
-      persistence.readRef.mockResolvedValue(existingSha);
-      persistence.showNode.mockResolvedValue(
-        `warp:patch\n\neg-kind: patch\neg-graph: test-graph\neg-writer: writer1\neg-lamport: 5\neg-patch-oid: ${existingPatchOid}\neg-schema: 2`
-      );
-
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1, // Constructor lamport is 1, but commit should use 6
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-        expectedParentSha: existingSha, // Race detection: expected parent matches current ref
-      } as any));
-
-      builder.addNode('x');
-      await builder.commit();
-
-      // Check the commit has lamport 6 (5 + 1)
-      const commitCall = persistence.commitNodeWithTree.mock.calls[0]![0];
-      const decoded = decodePatchMessage(commitCall.message);
-      expect(decoded.lamport).toBe(6);
-
-      // Parent should be the existing commit
-      expect(commitCall.parents).toEqual([existingSha]);
-    });
-
-    it('creates tree with patch.cbor blob', async () => {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      await builder.commit();
-
-      // Check writeTree was called with correct format
-      const treeCall = persistence.writeTree.mock.calls[0]![0];
-      expect(treeCall).toHaveLength(1);
-      expect(treeCall[0]).toMatch(/^100644 blob [a-f0-9]+\tpatch\.cbor$/);
-    });
-
-    it('writes patch blob with CBOR encoding', async () => {
-      const persistence = createMockPersistence();
-      const patchJournal = createPatchJournal(persistence);
-      const vv = VersionVector.empty();
-      vv.set('otherWriter', 3);
-
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal,
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: vv,
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x').setProperty('x', 'name', 'X');
-      await builder.commit();
-
-      // Decode the blob that was written
-      const blobData = persistence.writeBlob.mock.calls[0]![0];
-      const patch = decode(blobData) as any;
-
-      expect(patch.schema).toBe(2);
-      expect(patch.writer).toBe('writer1');
-      expect(patch.lamport).toBe(1);
-      expect(patch.ops).toHaveLength(2);
-      expect(patch.ops[0].type).toBe('NodeAdd');
-      expect(patch.ops[1].type).toBe('PropSet');
-      // Context should be serialized version vector
-      expect(patch.context).toBeDefined();
-    });
-
-    it('first commit has no parents', async () => {
-      const persistence = createMockPersistence();
-      // No existing ref
-      persistence.readRef.mockResolvedValue(null);
-
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      await builder.commit();
-
-      const commitCall = persistence.commitNodeWithTree.mock.calls[0]![0];
-      expect(commitCall.parents).toEqual([]);
-    });
-  });
-
-  describe('use-after-commit guard', () => {
-    /**
-     * Creates and commits a builder, returning the committed builder instance.
-     * @returns {Promise<{ builder: PatchBuilder, persistence: any }>}
-     */
-    async function createCommittedBuilder() {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-      builder.addNode('x');
-      await builder.commit();
-      return { builder, persistence };
-    }
-
-    it('throws after commit when calling addNode', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.addNode('y')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling removeNode', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.removeNode('x')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling addEdge', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.addEdge('a', 'b', 'rel')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling removeEdge', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.removeEdge('a', 'b', 'rel')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling setProperty', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.setProperty('x', 'name', 'Alice')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling setEdgeProperty', async () => {
-      const { builder } = await createCommittedBuilder();
-      expect(() => builder.setEdgeProperty('a', 'b', 'rel', 'since', '2026-01-01')).toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws after commit when calling attachContent', async () => {
-      const { builder, persistence } = await createCommittedBuilder();
-      const initialWriteBlobCalls = persistence.writeBlob.mock.calls.length;
-      await expect(builder.attachContent('x', 'payload')).rejects.toThrow('PatchBuilder already committed — create a new builder');
-      expect(persistence.writeBlob).toHaveBeenCalledTimes(initialWriteBlobCalls);
-    });
-
-    it('throws after commit when calling attachEdgeContent', async () => {
-      const { builder, persistence } = await createCommittedBuilder();
-      const initialWriteBlobCalls = persistence.writeBlob.mock.calls.length;
-      await expect(builder.attachEdgeContent('a', 'b', 'rel', 'payload')).rejects.toThrow('PatchBuilder already committed — create a new builder');
-      expect(persistence.writeBlob).toHaveBeenCalledTimes(initialWriteBlobCalls);
-    });
-
-    it('throws after commit when calling commit again', async () => {
-      const { builder } = await createCommittedBuilder();
-      await expect(builder.commit()).rejects.toThrow('PatchBuilder already committed — create a new builder');
-    });
-
-    it('throws during an in-flight commit when mutating or committing again', async () => {
-      let releaseReadRef: (value: string | null) => void = () => {};
-      const readRefPromise: Promise<string | null> = new Promise((resolve) => {
-        releaseReadRef = resolve;
-      });
-      const persistence = createMockPersistence();
-      persistence.readRef.mockImplementation(() => readRefPromise);
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      const pendingCommit = builder.commit();
-
-      expect(() => builder.addNode('y')).toThrow('PatchBuilder already committed — create a new builder');
-      await expect(builder.commit()).rejects.toThrow('PatchBuilder already committed — create a new builder');
-
-      releaseReadRef(null);
-      await expect(pendingCommit).resolves.toBe('c'.repeat(40));
-    });
-
-    it('allows reading ops/reads/writes/versionVector after commit', async () => {
-      const persistence = createMockPersistence();
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addEdge('user:alice', 'user:bob', 'follows');
-      await builder.commit();
-
-      const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-      expect(builder.ops).toHaveLength(1);
-      expect(builder.reads.has('user:alice')).toBe(true);
-      expect(builder.reads.has('user:bob')).toBe(true);
-      expect(builder.writes.has(edgeKey)).toBe(true);
-      expect(builder.versionVector.get('writer1')).toBe(1);
-    });
-
-    it('does NOT set _committed on failed commit (CAS ref advance throws)', async () => {
-      const persistence = createMockPersistence();
-      persistence.compareAndSwapRef.mockRejectedValueOnce(new Error('simulated compareAndSwapRef failure'));
-      const builder = new PatchBuilder(({
-        persistence,
-        patchJournal: createPatchJournal(persistence),
-        graphName: 'test-graph',
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: VersionVector.empty(),
-        getCurrentState: () => null,
-      } as any));
-
-      builder.addNode('x');
-      await expect(builder.commit()).rejects.toThrow('simulated compareAndSwapRef failure');
-      expect((builder as any)._committed).toBe(false);
-      expect((builder as any)._committing).toBe(false);
-    });
-  });
-
-  describe('reads/writes provenance tracking (HG/IO/1)', () => {
-    describe('NodeAdd', () => {
-      it('tracks nodeId as write', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addNode('user:alice');
-
-        expect(builder.writes.has('user:alice')).toBe(true);
-        expect(builder.reads.has('user:alice')).toBe(false);
-      });
-
-      it('includes writes in built patch', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addNode('user:alice').addNode('user:bob');
-
-        const patch = builder.build();
-        expect(patch.writes).toEqual(['user:alice', 'user:bob']);
-        expect(patch.reads).toBeUndefined(); // Empty reads omitted
-      });
-    });
-
-    describe('NodeRemove', () => {
-      it('tracks nodeId as read', () => {
-        const state = createMockState();
-        const existingDot = Dot.create('otherWriter', 5);
-        state.nodeAlive.add('user:alice', existingDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder.removeNode('user:alice');
-
-        expect(builder.reads.has('user:alice')).toBe(true);
-        expect(builder.writes.has('user:alice')).toBe(false);
-      });
-
-      it('includes reads in built patch', () => {
-        const state = createMockState();
-        const existingDot = Dot.create('otherWriter', 5);
-        state.nodeAlive.add('user:alice', existingDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder.removeNode('user:alice');
-
-        const patch = builder.build();
-        expect(patch.reads).toEqual(['user:alice']);
-        expect(patch.writes).toBeUndefined(); // Empty writes omitted
-      });
-    });
-
-    describe('EdgeAdd', () => {
-      it('tracks endpoint nodes as reads and edge key as write', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addEdge('user:alice', 'user:bob', 'follows');
-
-        // Reads both endpoint nodes
-        expect(builder.reads.has('user:alice')).toBe(true);
-        expect(builder.reads.has('user:bob')).toBe(true);
-
-        // Writes the edge key (encoded as from\0to\0label)
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-        expect(builder.writes.has(edgeKey)).toBe(true);
-      });
-
-      it('includes reads and writes in built patch', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addEdge('user:alice', 'user:bob', 'follows');
-
-        const patch = builder.build();
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-
-        expect(patch.reads).toEqual(['user:alice', 'user:bob']);
-        expect(patch.writes).toEqual([edgeKey]);
-      });
-    });
-
-    describe('EdgeRemove', () => {
-      it('tracks edge key as read', () => {
-        const state = createMockState();
-        const existingDot = Dot.create('otherWriter', 3);
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-        state.edgeAlive.add(edgeKey, existingDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder.removeEdge('user:alice', 'user:bob', 'follows');
-
-        expect(builder.reads.has(edgeKey)).toBe(true);
-        expect(builder.writes.has(edgeKey)).toBe(false);
-      });
-
-      it('includes reads in built patch', () => {
-        const state = createMockState();
-        const existingDot = Dot.create('otherWriter', 3);
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-        state.edgeAlive.add(edgeKey, existingDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder.removeEdge('user:alice', 'user:bob', 'follows');
-
-        const patch = builder.build();
-        expect(patch.reads).toEqual([edgeKey]);
-        expect(patch.writes).toBeUndefined();
-      });
-    });
-
-    describe('PropSet on node', () => {
-      it('tracks nodeId as both read and write', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.setProperty('user:alice', 'name', 'Alice');
-
-        expect(builder.reads.has('user:alice')).toBe(true);
-        expect(builder.writes.has('user:alice')).toBe(true);
-      });
-
-      it('includes in both reads and writes arrays', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.setProperty('user:alice', 'name', 'Alice');
-
-        const patch = builder.build();
-        expect(patch.reads).toEqual(['user:alice']);
-        expect(patch.writes).toEqual(['user:alice']);
-      });
-    });
-
-    describe('setEdgeProperty', () => {
-      it('tracks edge key as both read and write', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        // First add the edge, then set property
-        builder.addEdge('user:alice', 'user:bob', 'follows');
-        builder.setEdgeProperty('user:alice', 'user:bob', 'follows', 'since', '2025-01-01');
-
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-        expect(builder.reads.has(edgeKey)).toBe(true);
-        expect(builder.writes.has(edgeKey)).toBe(true);
-      });
-
-      it('includes edge key in built patch reads and writes', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addEdge('user:alice', 'user:bob', 'follows');
-        builder.setEdgeProperty('user:alice', 'user:bob', 'follows', 'since', '2025-01-01');
-
-        const patch = builder.build();
-        const edgeKey = encodeEdgeKey('user:alice', 'user:bob', 'follows');
-
-        // Reads includes alice, bob (from addEdge) and edgeKey (from setEdgeProperty)
-        expect(patch.reads).toContain('user:alice');
-        expect(patch.reads).toContain('user:bob');
-        expect(patch.reads).toContain(edgeKey);
-
-        // Writes includes edgeKey (from both addEdge and setEdgeProperty)
-        expect(patch.writes).toContain(edgeKey);
-      });
-    });
-
-    describe('complex patches', () => {
-      it('deduplicates reads and writes', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        // Add node twice via different operations should only appear once
-        builder.addNode('user:alice');
-        builder.setProperty('user:alice', 'name', 'Alice');
-        builder.setProperty('user:alice', 'email', 'alice@example.com');
-
-        const patch = builder.build();
-
-        // Should be deduplicated and sorted
-        expect(patch.writes).toEqual(['user:alice']);
-        expect(patch.reads).toEqual(['user:alice']);
-      });
-
-      it('sorts reads and writes deterministically', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        // Add nodes in non-alphabetical order
-        builder.addNode('user:zebra');
-        builder.addNode('user:alice');
-        builder.addNode('user:middle');
-
-        const patch = builder.build();
-
-        // Should be sorted
-        expect(patch.writes).toEqual(['user:alice', 'user:middle', 'user:zebra']);
-      });
-
-      it('handles mixed operations correctly', () => {
-        const state = createMockState();
-        // Pre-populate state with an existing node and edge
-        const nodeDot = Dot.create('writer0', 1);
-        const edgeDot = Dot.create('writer0', 2);
-        state.nodeAlive.add('user:existing', nodeDot);
-        const existingEdgeKey = encodeEdgeKey('user:existing', 'user:target', 'knows');
-        state.edgeAlive.add(existingEdgeKey, edgeDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder
-          .addNode('user:new')                           // writes: new
-          .addEdge('user:new', 'user:friend', 'likes')   // reads: new, friend; writes: edge
-          .removeNode('user:existing')                    // reads: existing
-          .removeEdge('user:existing', 'user:target', 'knows') // reads: existing edge
-          .setProperty('user:new', 'name', 'New');       // reads: new; writes: new
-
-        const patch = builder.build();
-        const newEdgeKey = encodeEdgeKey('user:new', 'user:friend', 'likes');
-
-        // Reads should include: user:new, user:friend (from addEdge),
-        // user:existing (from removeNode), existing edge key (from removeEdge)
-        expect(patch.reads).toContain('user:new');
-        expect(patch.reads).toContain('user:friend');
-        expect(patch.reads).toContain('user:existing');
-        expect(patch.reads).toContain(existingEdgeKey);
-
-        // Writes should include: user:new (addNode + setProperty), new edge key (addEdge)
-        expect(patch.writes).toContain('user:new');
-        expect(patch.writes).toContain(newEdgeKey);
-
-        // user:existing should NOT be in writes (only removed, not added)
-        expect(patch.writes).not.toContain('user:existing');
-      });
-    });
-
-    describe('backward compatibility', () => {
-      it('omits empty reads array from patch', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addNode('x'); // Only writes, no reads
-
-        const patch = builder.build();
-        expect(patch.writes).toEqual(['x']);
-        expect(patch.reads).toBeUndefined();
-      });
-
-      it('omits empty writes array from patch', () => {
-        const state = createMockState();
-        const existingDot = Dot.create('otherWriter', 5);
-        state.nodeAlive.add('x', existingDot);
-
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 2,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => state,
-        } as any));
-
-        builder.removeNode('x'); // Only reads, no writes
-
-        const patch = builder.build();
-        expect(patch.reads).toEqual(['x']);
-        expect(patch.writes).toBeUndefined();
-      });
-
-      it('handles patch with no ops gracefully (builds but cannot commit)', () => {
-        const builder = new PatchBuilder(({
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        // Build empty patch (no ops)
-        const patch = builder.build();
-        expect(patch.reads).toBeUndefined();
-        expect(patch.writes).toBeUndefined();
-        expect(patch.ops).toEqual([]);
-      });
-    });
-
-    describe('commit() includes reads/writes', () => {
-      it('committed patch includes reads/writes arrays', async () => {
-        const persistence = createMockPersistence();
-        const patchJournal = createPatchJournal(persistence);
-        const builder = new PatchBuilder(({
-          persistence,
-          patchJournal,
-          graphName: 'test-graph',
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addNode('user:alice').setProperty('user:alice', 'name', 'Alice');
-        await builder.commit();
-
-        // Decode the CBOR blob that was written
-        const blobData = persistence.writeBlob.mock.calls[0]![0];
-        const patch = decode(blobData) as any;
-
-        expect(patch.reads).toEqual(['user:alice']);
-        expect(patch.writes).toEqual(['user:alice']);
-      });
-
-      it('committed patch omits empty reads array', async () => {
-        const persistence = createMockPersistence();
-        const patchJournal = createPatchJournal(persistence);
-        const builder = new PatchBuilder(({
-          persistence,
-          patchJournal,
-          graphName: 'test-graph',
-          writerId: 'writer1',
-          lamport: 1,
-          versionVector: VersionVector.empty(),
-          getCurrentState: () => null,
-        } as any));
-
-        builder.addNode('x'); // Only writes, no reads
-
-        await builder.commit();
-
-        const blobData = persistence.writeBlob.mock.calls[0]![0];
-        const patch = decode(blobData) as any;
-
-        expect(patch.writes).toEqual(['x']);
-        expect(patch.reads).toBeUndefined();
-      });
-    });
-  });
-
-  describe('removeNode / removeEdge without materialized state', () => {
-    it('removeNode throws PatchError when state is null', () => {
-      const vv = VersionVector.empty();
-      const builder = new PatchBuilder(({
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: vv,
-        getCurrentState: () => null,
-      } as any));
-
-      expect(() => builder.removeNode('alice')).toThrow(PatchError);
-    });
-
-    it('removeNode error has code E_PATCH_NO_STATE', () => {
-      const vv = VersionVector.empty();
-      const builder = new PatchBuilder(({
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: vv,
-        getCurrentState: () => null,
-      } as any));
-
-      expect(() => builder.removeNode('alice')).toThrow(
-        expect.objectContaining({ code: 'E_PATCH_NO_STATE' }),
-      );
-    });
-
-    it('removeEdge throws PatchError when state is null', () => {
-      const vv = VersionVector.empty();
-      const builder = new PatchBuilder(({
-        writerId: 'writer1',
-        lamport: 1,
-        versionVector: vv,
-        getCurrentState: () => null,
-      } as any));
-
-      expect(() => builder.removeEdge('a', 'b', 'knows')).toThrow(PatchError);
-    });
-
-    it('removeNode works when state is available', () => {
-      const state = createMockState();
-      state.nodeAlive.add('alice', Dot.create('writer1', 1));
-
-      const vv = VersionVector.empty();
-      const builder = new PatchBuilder(({
-        writerId: 'writer1',
-        lamport: 2,
-        versionVector: vv,
-        getCurrentState: () => state,
-      } as any));
-
-      builder.removeNode('alice');
-      const patch = builder.build();
-      expect(patch.ops).toHaveLength(1);
-      const op = patch.ops[0] as any;
-      expect(op.type).toBe('NodeRemove');
-      expect(op.observedDots.length).toBeGreaterThan(0);
-    });
-  });
 });
